+++ examples/giano/src/@adamquadmon/gatsby-theme-lucifero/hooks/use-navItems.js
@@ -1,11 +1,14 @@
 import { useTranslation } from 'gatsby-plugin-react-i18next'
+import { useContent } from '../../../hooks/use-content'
 
-export const useNavItems = () => {
+export const useNavItems = (language) => {
   const { t } = useTranslation()
+  const content = useContent(language)
   return [
     {
       label: t('navLink1Label'),
       href: t('navLink1Href'),
+      children: content,
     },
   ]
 }