+++ themes/gatsby-theme-lucifero/src/components/NavBar/NavItemSub.js
@@ -9,14 +9,14 @@ const NavItemSub = ({ variant, item, ...rest }) => {
   const styles = useStyleConfig('NavItemSub', { variant })
   const { body, frontmatter } = item
   const { cover, slug, title } = frontmatter
-  const coverImage = getImage(cover)
+  const coverImage = cover && getImage(cover)
   return (
     <Box __css={styles} {...rest}>
       <Link to={slug}>
         <Heading as="h4" size="md" className="item_header">
           {title}
         </Heading>
-        <Image as={GatsbyImage} image={coverImage} alt={title} />
+        {cover && <Image as={GatsbyImage} image={coverImage} alt={title} />}
         <Box className="item_content">
           <MDXRenderer>{body}</MDXRenderer>
         </Box>