+++ examples/giano/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.0.1
+
+Add sub content to test NavItemSub
+
 ## 1.0.0
 
 First release